Removes `AbstractManifest`
